+++ RSPiX/Src/GREEN/3D/zbuffer.cpp
@@ -54,9 +54,9 @@ int16_t RZBuffer::Create(int16_t sW,int16_t sH)
 
 	m_sW = sW;
 	m_sH = sH;
-	m_lP = long((sW+3)&~3); // keep for emergencies!
+	m_lP = int32_t((sW+3)&~3); // keep for emergencies!
 
-	long lSize = m_lP * m_sH * sizeof(int16_t);
+	int32_t lSize = m_lP * m_sH * sizeof(int16_t);
 	m_pBuf = (int16_t*) malloc(lSize);
 	// you then may clear it, buddy!
 
@@ -102,11 +102,11 @@ void RZBuffer::Clear(int16_t sVal)
 	BigWord.p1 = BigWord.p2 = BigWord.p3 = BigWord.p4 = sVal;
 
 	//2) Do the copy
-	long lWordP = m_lP >> 2; // 4 * 16 = 64
+	int32_t lWordP = m_lP >> 2; // 4 * 16 = 64
 	U64* pWord = (U64*) m_pBuf;
 
-	long lWordLen = lWordP * m_sH;
-	for (long i=0;i < lWordLen; i++) *pWord++ = BigWord.word;
+	int32_t lWordLen = lWordP * m_sH;
+	for (int32_t i=0;i < lWordLen; i++) *pWord++ = BigWord.word;
 
 	}
 
@@ -116,7 +116,7 @@ void RZBuffer::TestHeight(RImage* pimDst,int16_t sDepth,
 				int16_t sX,int16_t sY,int16_t sW,int16_t sH)
 	{
 	int16_t i,j;
-	long lP = pimDst->m_lPitch;
+	int32_t lP = pimDst->m_lPitch;
 
 	UCHAR *pDst,*pDstLine = pimDst->m_pData + lP * sY + sX;
 	int16_t *pZB,*pZLine = m_pBuf + m_lP * sY + sX;
@@ -136,7 +136,7 @@ void RZBuffer::TestHeight(RImage* pimDst,int16_t sDepth,
 void RZBuffer::Dump(RImage* pimDst,int16_t sX,int16_t sY,int16_t sW,int16_t sH,UCHAR* pZCol)
 	{
 	int16_t i,j;
-	long lP = pimDst->m_lPitch;
+	int32_t lP = pimDst->m_lPitch;
 
 	UCHAR *pDst,*pDstLine = pimDst->m_pData + lP * sY + sX;
 	int16_t *pZLine = m_pBuf + m_lP * sY + sX;
